requirements.txt
# ============================================
# CORE LANGCHAIN (Compatible versions)
# ============================================
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langsmith==0.1.129

# ============================================
# MULTI-AGENT & LLM
# ============================================
langchain-groq==0.1.9
langgraph==0.2.0
groq==0.9.0

# ============================================
# EMBEDDINGS
# ============================================
sentence-transformers==2.3.1

# ============================================
# VECTOR DATABASE
# ============================================
faiss-cpu==1.7.4

# ============================================
# PDF PROCESSING
# ============================================
PyMuPDF==1.23.8
pypdf==3.17.4

# ============================================
# WEB FRAMEWORKS
# ============================================
fastapi==0.108.0
uvicorn[standard]==0.25.0
streamlit==1.29.0

# ============================================
# WEB SEARCH
# ============================================
tavily-python==0.2.8
duckduckgo-search==4.1.0

# ============================================
# UTILITIES (Fixed pydantic version)
# ============================================
python-dotenv==1.0.0
pydantic==2.7.4
faker==22.0.0
rich==13.7.0

# ============================================
# DATA PROCESSING
# ============================================
pandas==2.1.4
numpy==1.26.3

# ============================================
# HTTP & ASYNC
# ============================================
aiohttp==3.9.1
requests==2.31.0